Fraction extends number

+++ src/com/qmul/matrix/Matrix2x2.java
@@ -32,7 +32,7 @@ public class Matrix2x2 extends Matrix{
     public Matrix2x2 inverse(){
         Fraction[][] invertedMatrix = new Fraction[2][2];
         Fraction det = this.determinant();
-        if(det.toDecimal()!=0){
+        if(det.intValue()!=0){
             invertedMatrix[0][0] = det.multiply(this.get(1, 1));
             invertedMatrix[0][1] = det.multiply(this.get(0,1)).negate();
             invertedMatrix[1][0] = det.multiply(this.get(1,0));
